proxy: Pass addressing info for response logs correctly
Pass source vs. destination addressing info in access logs correctly
instead of relying Hubble to swap them around for responses. This has the
benefit that also cilium monitor output has correct address and security
identity for L7 responses.

Note that this changes cilium proxy output that used to incorrectly show
security identities as "A->B" for both requests and responses. Now these
are formatted as "A->B" and "B->A" for requests and responses,
respectively.

Fixes: #9558
Signed-off-by: Jarno Rajahalme <jarno@isovalent.com>